+++ src/main/java/enginetown/Scene.java
@@ -4,5 +4,5 @@ public abstract class Scene {
     public Scene() {
     }
 
-    public abstract void update(float dt);
+    public abstract void update(double dt);
 }
