+++ src/main/java/enginetown/Window.java
@@ -110,9 +110,9 @@ public class Window {
     }
 
     public void loop() {
-        float beginTime = Time.getTime();
-        float endTime;
-        float dt = -1.0f;
+        double beginTime = Time.getTime();
+        double endTime;
+        double dt = -1.0;
 
         while (!glfwWindowShouldClose(glfwWindow)) {
             // Poll events
